requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.2
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.1.1
flake8-quotes==3.4.0
flake8-variables-names==0.0.6
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.5
mccabe==0.7.0
pep8-naming==0.13.2
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
psycopg==3.1.12
psycopg-binary==3.1.12
